refactor TransitionResult into an enum
this makes the code much simpler to use, to read and less error
prone.

this commit fixes the tests as well.
`is_quit` is implemented in `test` to make the tests easier to write,
i.e. by writing `transition_result.is_quit()` instead of matching
for `Quit` or `Return` each time.