upgrade the plugin to 0.84.1 (#22)
I realize that you might not accept this, but I had a few minutes to try
and upgrade this to the latest nushell main branch. Feel free to change
anything you want in this PR, no hurt feelings here.

This is done with the least amount of effort, but it compiles and tries
to run on Windows. What I would recommend, if you choose to land this,
is spending some time looking at the records in the project, because
there are better ways to implement them in 0.84.1. That isn't saying
your code isn't good. That's saying that what I did, I did with the
least amount of effort and didn't try to optimize anything.

Anyway, this should give you a head-start with porting your plugin to
the latest version of nushell.

---------

Co-authored-by: amtoine <stevan.antoine@gmail.com>