add support from table / record transposition (#28)
this PR adds the ability to tranpose the current view, if it's a table
or a record, just as the built-in `transpose` command does.

the default binding is `t`.

the transposition is an *involution*, i.e. it is its self inverse,
meaning that applying the tranposition twice give you back the original
data.