Fix TUI channel panics (#63)
Fix panics when sending to the TUI channel.

Because failing to send just means the other end hung up, we don't
really need to do anything special. It's totally expected that the other
side will hang up eventually. It's best for us to just break out of the
loop.

This also updates nushell to a revision I have right now that fixes
local socket, but this should probably be changed on release.

---------

Co-authored-by: amtoine <stevan.antoine@gmail.com>